Extract Multiquestion logic into a separate class
Make Multiquestion a separate class and remove all logic related
to `self.questions` from `Question`. This removes if statements
from the `Quesiton` methods and makes sure that for each qustion
with `type: multiquestion` we end up with a `Multiquestion` instance
by adding the new class to the `QUESTION_TYPES` dictionary.

Most of the code is unchanged, the most tricky part of the refactoring
is probably the multiple inheritance we need to use for the summary
class. Since `MultiquestionSummary` needs both `QuestionSummary`
methods and `Multiquestion` methods instead of base `Question` ones,
we need to create the diamond inheritance that Python resolves to
`Multiquestion Summary` - `QuestionSummary` - `Multiquestion` -
`Question` method resolution order, which gives us exactly what we
want.