Keep the .form_fields order in .required_form_fields
Using set() doesn't guarantee a consistent order of form fields,
which means we need to make sure tests are using `sorted()` when
making assertions on the method return value. Keeping the order
removes the need for this and since the number of fields per
question is low there should be no performance difference.