Filter ContentSection description by lot
When provided with a description which is a dictionary of lot keys,
we are able to filter this dictionary by lot. If a lot key is not
found then we can fall back to the `default` key. If no `default`
key exists in the description, then we raise an error.

Introduced description as a `@property` so we can raise an error if
we try to access the description but it has not yet been filtered
to it's chosen string.

Existing behaviour of parsing a description which is a string or is
`None` is not changed.