Questions filter themselves
Since we're going to be templating question fields, we're going
to need to be passing data into the questions themselves so that
they can decide what to do with it.

This commit doesn't allow any question fields to be templated, but
it *does* delegate the decision around filtering questions down
to the Question itself.  Now a section will call `.filter` on its
questions and see what comes back.