Move from templating fields in .filter to TemplateField
Adds utils.TemplateField class that parses the field value as
Jinja2 template. Creating the template from string is an expensive
operation, so we want to do this once when the content file is loaded
for the first time, which is why we're changing ContentLoader to replace
templatable fields with TemplateField objects.

TemplateField stores the original template string to make it possible to
compare two instances in tests.