Render TemplateFields on access in ContentSection
We used to prevent access to templatable fields before the
`/filter` method had been called (by checking the `_filtered`
flag).  Now, instead of rendering the fields when `.filter` is
called, we just capture the context and then render the fields
when the attribute is accessed.

This means that we don't need to iterate over/overwrite field
values in the `.filter` method since `__getattribute__` can
check the type when fields are accessed.
We can also avoid restricting access to templatable fields that
don't take any context variables since the templates render without
raising an error.