Run template fields through Markdown first
At the moment, some of the content fields are processed with a
markdown filter to make writing question hints and descriptions
easier.

Since we're introducing Jinja to replay user input in question and
section data we need to move the markdown processing before Jinja,
since otherwise we might render user input as markdown.

We're only applying markdown to multiline fields since that's where
it's used by the content repo and we want to avoid wrapping single
line strings with paragraph tags.

This also introduces a number of issues with Markdown recognising
some of the characters in Jinja tags (eg '<', '*', '_') and causing
template errors by replacing them with HTML. Our main concern is that
we raise these errors early since they don't seem to prevent us from
doing common things we want to do with Jinja.