Test that XSS vulnerabilites are mitigated
To prevent cross-site scripting attacks, we want to make sure that
HTML or jinja templating that would be passed in through the
context don't end up as valid HTML once the template is rendered.

Tests make sure that HTML tags are escaped and that jinja tags are
printed literally instead of being interpreted.