Create `DynamicList` question
A DynamicList question is a type of multiquestion whose
inner questions are not generated until runtime.
We are trying to create a list of nice-to-have questions
but we only know what the question is going to be once
we've seen the brief itself.
Thus, the `.filter()` method of a DynamicList question
will create new questions with the correctquestion text,
id, and potential followup value.