Return correct error messages to template
Errors for dynamic_list questions need to be linked to `the field
that is failing the validation: either the primary question or a
follow-up question (depending on the value of the primary question).
requirement, we then have to have a value for 'evidence'.

Accordingly, we're expecing to receive an error dict with
a `field` key back from the API that will tell us where to
stick the validation message.