Rewrite test examples to be clearer and more realistic
Dyanmic list test examples used keys with very similar names and
readability was hard. We have also changed the example to be more
realistic i.e. the first question is a boolean (as opposed to
a text field as before) and the follow up question is a text field.
This mirrors the example of providing evidence for the new
supplier flow.