Remove followup answers for False answer to dynamic list questions
Our form data may include followup answers which are non empty
(if the user says yes to the question, adds followup answer and
then changes their mind and says no to the original question).
This would mean that `False` would be sent with the follow up
answer in the form data. We wish to strip out the follow up answer
in this case when using `get_data` as it is redundant information.