Fix dynamic list filtering for schema generator
Schema generation calls `.filter` to create separate manifests for
separate framework lots, so that it can generate a schema for each
lot.

With the addition of DynamicList questions, filter now requires
dynamic data to generate the actual list of questions specific to
the brief, but that's not something we want to happen when listing
the questions for the schema generation. Schemas should contain
the original generic nested questions, so we need a separate way to
filter the manifests without doing the dynamic question replacements.

In order to achieve this, I'm adding a `static` flag to `.filter`,
which when set to True will make it skip the "dynamic" parts of
the filtering logic.