Add support for the followup questions to the multiquestion
Supports new followup YAML format.

Makes a separate _drop_followups method for the Multiquestions and
adds a call to it from `.get_data`.

It extends the method to support multi-input followup questions (eg
it should correctly drop all fields for a pricing followup), but it
still assumes that the original question data is stored in question
id key (which is true for most questions apart from multiquestions
and pricing, so they cannot have followups at the moment).