Make drop_followups return None values for top-level data fields
For nested questions (eg questions insidea a dynamic list array) we remove the question field
completely, since the top-level data key will be replaced anyway.

For multiquestions that are serialized to separate top-level keys we set the follow-up value
to `None`, so that it's replaced if the question was previously answered with a follow-up.