Add support for checkboxes questions with followups
Since checkboxes return a list of strings from `.get_data`, drop
followups needs to support two cases:

* check if a string/boolean value is in a list of followup values
  (for boolean and radios questions)
* check that any of the strings in the checkboxes question answer
  are in the list of followup values

Ideally, this logic should be encapsulated in separate question types,
so that eg boolean questions know to check for the first case and
checkboxes questions only check for the second one.

But this requires us to either duplicate the logic in each class just
to change the condition or pass a lot of arguments back to the question
(including the list of values for the current follow-up and the full
data dict) to only delegate the "in" check. In addition, radios and
checkboxes share the same question class at the moment.

So instead I'm extending the comparison logic so that it supports
both cases.