Add new filter_value property to question summaries
Some questions have additional "filter_label"s for their options but there's currenly no way to get these from the question summary to display on a page.

This adds a new `filter_value` property that will return the `filter_label` if it exists for the selected option, or just the current `value` if it doesn't.