Permissive validation for pounds and optionally pence
If `decimal_place_restriction` is True, then we are wishing to
restrict our input to pounds and optionally pence. So we only
store things in the database that look like 'pretty' monetary
values, we add permissive validation to format values that
don't look 'pretty' but do represent valid pounds or pounds with
pence. So even if a user enters '99.0', we will transform it to
'99.00' for which it can then pass validation on the API rather
than give them an error message and force them to correct it.