Load arbitrary framework metadata
 ## Summary
We need to be able to copy services from one framework to another during
applications to save suppliers time. In order to do this, we need to
define which previous framework has services that are eligible for
copying, and which questions are suitable. This information appears to
be data related to the framework, so we are storing it in dm-frameworks
as metadata about a given framework (in this case, G-Cloud 10). In order
for a frontend to have that information, we need to be able to load it.
The `messages` folder of dm-frameworks stores files that contain
key-value pairs, and this is the same kind of information format we want
for our metadata, so this creates a stripped-down version of the same
which works on the `metadata` folder of a framework. A ContentMetadata
instance does not parse the data it reads and therefore does not support
templated fields; this seems like a reasonable conclusion to make as we
are simply storing static data that describes how the framework
interacts with others.