Use snapshot testing to test govuk_input
To try and keep the tests of the functions in govuk_frontend both
comprehensive and readable, I've used snapshot testing with `syrupy` to
move the expected values into a separate file.

The benefit of this is that the tests are more readable and easier to
update (just run `venv/bin/pytest --update-snapshots`), and if the test
fails we get very pretty diffs highlighting the change in question.

The downside is that the expected value is now outside of the test file,
which means a bit less transparency.