Change Question.question_advice to be always Markdown
We're having issues where the question advice is sometimes wrapped in
HTML `p` tags, and sometimes isn't.

Ideally it would be consistent, so this commit changes the behaviour of
the content loader to always treat the `question_advice` field as
Markdown. This means it will always be HTML `<p>`s!