Move implementation of href property into govuk_frontend module
We want to couple the implementation of the `Question.href` logic more
closely with other govuk-frontend code, so the definition of the
implementation should belong in `dmcontent.govuk_frontend`.

This does unfortunately mean that we have an import cycle between
questions.py and govuk_frontend.py if we want to have type annotations.
We work around this by following the instructions from mypy [[1]]. The
upside is that the inter-dependencies between questions.py and
govuk_frontend.py should be more explicit and less two-way now.

[1]: https://mypy.readthedocs.io/en/stable/common_issues.html#import-cycles