Add support for pricing without minima
In DOS 5, we will only have the maximum. Update the summary so this gets handled correctly. Add tests to verify. Add in some type hints for fun.

I am assuming that nothing that currently exists relies on the previous behaviour. As far as I can tell, form verification ensures that the minimum is provided where it is needed.

Also drop support for min price being an empty string, which would output nonsense (e.g. "£ per unit per day"). The implementation allows the min/max to be 0. This would be unusual, but not an impossible price for someone to offer a service to govenment.