Add suffix and prefix arguments to convert_to_number
We want to be able to gracefully handle the case where users input units
into a form, following Design System guidance [[1]].

This commit adds two new keyword-only arguments to `convert_to_number()`,
`prefix` and `suffix`. If a suffix or a prefix (or both) is provided,
and the user has included the same suffix or/and prefix in their data,
then that prefix and/or suffix will be removed.

The code to remove prefixes or suffixes is copied from [PEP-616][2], one
day we might be able to replace it with the `removeprefix()` and
`removesuffix()` methods.

[1]: https://design-system.service.gov.uk/components/text-input/#prefixes-and-suffixes
[2]: https://www.python.org/dev/peps/pep-0616/