Choose whether to use 'a' or 'an'
Previously, we always used 'a'. This doesn't work for things like 'a instance', and a supplier has noticed this. Fixing this properly will be a bit involved. So add a heuristic that should work most of the time.

If this is not enough, we should just use inflect rather than trying to write code that understands english better than this.