Add support for followup questions in multiquestions
We want to be able to handle multiquestions that contain questoins with
followups, like DOS5's [niceToHaveRequirements and yesNo questions][1].
The Design System calls this [conditionally revealing content][2].

This commit adds logic to `dm_multiquestion()` to handle these
questions.

We put the logic here rather than in say `govukRadios()` because a) the
way followups currently work is a little confusing, and b) it would be
better in the long term to re-evaluate how we lay out these questions
for the user and define them in the framework manifests. For now this
commit gives us some form of backwards compatibility to unblock
migration to govuk-frontend.

[1]: https://github.com/alphagov/digitalmarketplace-frameworks/blob/master/frameworks/digital-outcomes-and-specialists-5/questions/brief-responses/niceToHaveRequirements.yml
[2]: https://design-system.service.gov.uk/components/radios/#conditionally-revealing-content