Create a lazy mutable mapping
We want the ability to load content lazily to improve app startup times. Create a lazy dictionary to support this.

It implements all the required methods, and has some tests to check the lazy functionality.

It works by checking whether a value is callable when it is requested. If it is callable, then it is called. The result is stored in the value, and also returned.