Allow lazy loading of manifests
Manifests are slow to load - they take ~1 second each. This contributes to very slow app startup times - ~20-30 seconds for supplier-frontend.

Many of the manifests are for old frameworks that are requested very rarely or not at all. Thus, we can lazily load these old frameworks to give us a speedup on app startup without significantly slowing down real users.

See https://github.com/alphagov/digitalmarketplace-supplier-frontend/pull/1348 for the experimental version of this I used for testing. Based on the numbers in that PR, I think this will save developers a few minutes per day of waiting for tests, which will add up significantly over the course of a year.

Note that the LazyDict used is not thread safe. This is not a problem. The worst that could happen is that two threads simultaneously load the manifest. This would result in some wasted work. However, loading a manifest is idempotent, so there would be no other harmful effects.