Add support for file upload questions
We want to use the `govukFileUpload` component[[1]]
for our questions with `type: upload`[[2]]. Add a branch to
`render_question` that generates parameters suitable for this
component.

The old frontend toolkit showed the url of any previously
uploaded file. Add this to the hint text, after any question
advice or hints.

https://trello.com/c/7Hw9pH6X/855-2-add-type-upload-to-content-loader-renderquestion

[1]: https://design-system.service.gov.uk/components/file-upload/
[2]: https://github.com/alphagov/digitalmarketplace-frameworks/blob/a691598f98aa4a21c2615bdc07437898abe1591e/frameworks/digital-outcomes-and-specialists-5/questions/declaration/modernSlaveryStatement.yml