Fix type hints for multiquestions
For multiquestions, `from_question` returns a list of Markup and dicts. Create a type alias for this.
When rendering questions from a multiquestion, `from_question` will always return a dict because we don't have rnested multiquestions. Cast the return type as a dict so we know we can index it.

Also move the exception for a question type we don't know about from `render_question` to `from_question`. This allows us to simplify the type signature of `from_question` as it can no longer return None.
No apps call `from_question` directly, so this won't change behaviour.