Fix type hints for Question.filter
We now ensure that when a new question is created with `ContentQuestion` it has the same type as the question it was created from.
We've added a test to one of the `QuestionTest` subclasses - we can't add it to the base class because we need a Question of a different type.