Mark this package as having type hints
As per [PEP 561](https://mypy.readthedocs.io/en/stable/installed_packages.html#making-pep-561-compatible-packages), we need to let type checkers know that this package has type hints. We do this by adding an empty `py.typed` file, and adding it to `setup.py`.

mypy will now recognise that this package is typed. So it will check the types when type checking things that depend on dmcontent.