Add default maxItems for List Input
dmListInput requires maxItems to be defined in order to render the correct number of
inputs.

On Supplier Frontend, we have some questions that do not define `number_of_items` since
the Frontend Toolkit forms macro passed a default of 10 into the old List Entry component.

Rather than hunting down all the questions that do this and adding `number_of_items` to
each, we can pass the same default here.