Add get_question_as_section_containing_itself
We use the digitalmarketplace-content-loader `render_question` method  to render forms across our frontend apps.

This method expects a single question as input, but in some cases we're using `get_question_as_section` to convert
Questions to ContentSections. This has the side-effect of removing information about the question itself. For
example, we lose the `type` property.

To deal with this, we've created a new method which retains this question information so that we can access it normally.