Use appropriate label and legend classes
We want to have a visual heirarchy between questions and their conditional content. Additionally, we need to account
for (buggy) cases where some labels display as normal text and others display with appropriate font size styling.

Fieldset legends default to `--m` sizing, and this can be overridden. This also fixes a bug where this default was not
overridden if the legend was set to the page heading. Labels do not have a default.

In the case of multiquestions, we use `--s` sizing for each of the subquestions.

This change allows the template calling `render_question` to define these classes.