Fix bug when generating list HTML
Fix a bug where invalid HTML was generated when a HTML list was put in
one of a question's Markdown fields (eg. `essentialRequirementsMet`[1])

This was because the markdown preprocessor (`markdown.preprocessors.HtmlBlockPreprocessor`)
replaces raw HTML content with a placeholder value wrapped in an empty
`<p></p>`. There is a postprocessor (`markdown.postprocessors.RawHtmlPostprocessor`)
that finds the placeholders and removes the empty `<p></p>`,
however we have a custom [`Treeprocessor`](https://github.com/alphagov/digitalmarketplace-content-loader/blob/main/dmcontent/markdown.py#L8)
which added on the `govuk-body` class to these empty `<p>` tags and
meant that they weren't removed by the post processor.

This meant that we had list elements wrapped in `<p>` tags in the
output HTML which isn't valid HTML.

Add a check to `dmcontent.markdown._ClassAddingTreeprocessor` that
ensures it doesn't add classes to these placeholder elements.

[1]:https://github.com/alphagov/digitalmarketplace-frameworks/blob/master/frameworks/digital-outcomes-and-specialists-5/questions/brief-responses/essentialRequirementsMet.yml