Fix bug when number field value is 0
In some cases, we may pass 0 in a number field and expect it to be a valid answer.

This doesn't display as 0 is falsy. We can fix this by stringifying ints in a govuk_input,
since in the generated HTML, it's expecting a string anyway, since govuk_frontend doesn't
use inputs of type number.