Add integer check to Question._should_be_shown()
Up till now, we've exclusively used the `depends` property for filtering by lot, so we're always checking an array of
strings for one that matches the filter.

However, some questions are dependent on others having a specific answer, but are not direct conditionals. Currently,
this only applies to `socialValueCriteriaOutcomes`, which requires that `socialWeighting` be zero, or between 10 and 20.

This commit adds the ability to check one or several number ranges. It could also be rejigged in the future to allow
for checking multiple properties on any question type, but for now only applies to numbers.