Add support for Flask 2.0
We want to start supporting Flask 2.0 because we think it will bring performance improvements. We still need to support Flask 1.x to avoid breaking the users of this library.

Note there are deprecation warnings with Flask 2.0 that will become errors in 2.1. We cannot fix them (and thus add support for 2.1) until we drop support for 1.0